requirements.txt
langchain>=0.2.0
langgraph>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
faiss-cpu>=1.7.4
tiktoken>=0.7.0
python-dotenv>=1.0.0
typing_extensions>=4.9.0
